+++ src/tcode_api/servicer/client.py
@@ -2,8 +2,10 @@
 
 import importlib
 import logging
+import math
 import os
 import time
+from datetime import datetime
 from itertools import batched
 from typing import Any
 
@@ -23,7 +25,7 @@ from tcode_api.servicer.servicer_api import (
     SerialNumberLookupResponse,
 )
 from tcode_api.types import Matrix
-from tcode_api.utilities import generate_id, mm, rad
+from tcode_api.utilities import format_seconds, generate_id, mm, rad
 
 _logger = logging.getLogger(__name__)
 
@@ -234,10 +236,7 @@ class TCodeServicerClient:
         :returns: A list of reports on the attempted scheduling, in the same order as the commands were given. See :py:class:`ScheduleCommandResponse` for details.
         """
         tcode_api_version = tcode_api_version or self.tcode_api_version
-        data = [
-            ScheduleCommandRequest(command_id=id, command=command.model_dump()).model_dump()
-            for id, command in commands
-        ]
+        data = [[id, command.model_dump()] for id, command in commands]
         rsp = requests.put(
             f"{self.servicer_url}/{tcode_api_version}/schedule_commands",
             json=data,
@@ -325,7 +324,7 @@ class TCodeServicerClient:
         )
         rsp.raise_for_status()
 
-    def execute_run_loop(self) -> None:
+    def execute_run_loop(self, display_progress: bool = True) -> None:
         """Run a blocking loop that monitors the servicer's status and exits when the current
             script is complete or an error occurs.
 
@@ -333,6 +332,7 @@ class TCodeServicerClient:
             and exit cleanly.
         """
         self.set_run_state(True)
+        time_start = datetime.now()
         while True:
             try:
                 time.sleep(0.1)
@@ -340,6 +340,8 @@ class TCodeServicerClient:
 
                 if status.operation_count == 0:
                     self.set_run_state(False)
+                    if display_progress:
+                        print("\nTCode script successfully executed")
                     return
 
                 if not status.result.success:
@@ -350,12 +352,22 @@ class TCodeServicerClient:
                     _logger.fatal(msg)
                     self.set_run_state(False)
                     return
-
+                elapsed = (datetime.now() - time_start).total_seconds()
+                if display_progress:
+                    print(
+                        f"\rTCode executing: {status.operation_count:> 6} commands remaining. Time elapsed: {format_seconds(elapsed)}",
+                        end="",
+                        flush=True,
+                    )
             except KeyboardInterrupt:
+                print()  # newline
                 self.set_run_state(False)
                 self.clear_tcode_resolution()
                 self.clear_labware()
                 return
+            except:
+                print()  # newline
+                raise
 
     def run_script(
         self,
@@ -363,6 +375,7 @@ class TCodeServicerClient:
         clean_environment: bool = True,
         batch_process: bool = False,
         enable_socketio_user_input: bool = True,
+        display_progress: bool = True,
     ) -> None:
         """Schedule and execute a TCode script on the fleet, starting from an empty state.
 
@@ -402,6 +415,14 @@ class TCodeServicerClient:
                     raise RuntimeError(msg)
         else:
             batch_size = 100
+            scheduled_count = 0
+            command_size_digit_count = math.floor(math.log10(len(script.commands))) + 1
+            if display_progress:
+                print(
+                    f"\rTCode scheduling {scheduled_count:> {command_size_digit_count}}/{len(script.commands)}",
+                    end="",
+                    flush=True,
+                )
 
             for batch_index, command_batch in enumerate(batched(script.commands, batch_size)):
                 commands_list = list(command_batch)
@@ -420,10 +441,17 @@ class TCodeServicerClient:
                         if res.result.details is not None:
                             for line in res.result.details.get("traceback", "").split("\n"):
                                 _logger.debug(line)
+                        print()  # new line
                         raise RuntimeError(msg)
 
+                if display_progress:
+                    print(
+                        f"\rTCode scheduling {scheduled_count:> {command_size_digit_count}}/{len(script.commands)}",
+                        end="",
+                    )
+        print()  # new line
         try:
-            self.execute_run_loop()
+            self.execute_run_loop(display_progress)
         finally:
             try:
                 if sio is not None: